% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrapidmarkovFunctions.R
\name{runmarkov}
\alias{runmarkov}
\title{Calculate Markov trace}
\usage{
runmarkov(
  transitionsMatrix,
  startingStates,
  cycles = NULL,
  stateNames = NULL,
  quietly = FALSE
)
}
\arguments{
\item{transitionsMatrix}{2d matrix or 3d array of transition probabilities.}

\item{startingStates}{Vector of starting states.}

\item{cycles}{Number of cycles to run (only need to specify if
stationary matrix)}

\item{stateNames}{Vector of state names (optional)}

\item{quietly}{Boolean, if TRUE will print summary data about the model}
}
\value{
n*s matrix containing proportion of cohort in each state each
  cycle, where n=number of cycles, s=number of states.
}
\description{
Returns Markov trace as matrix of proportion of cohort in each state at each
  time point.
}
\examples{
# Three state stationary model (i.e. probabilities don't vary with time),
# ten cycles

states <- c("alive", "progressive", "dead")
startingStates <- c(1, 0, 0)
cycles <- 10
transitionsMatrix <- matrix(data = c(0.8, 0.1, 0.1,
                                     0.0, 0.7, 0.3,
                                     0.0, 0.0, 1.0),
                                     ncol = 3, byrow = TRUE)
runmarkov(transitionsMatrix, startingStates, cycles, states)


# Three-state non-stationary model (probabilities vary with time), six
# cycles.

states <- c("alive", "progressive", "dead")
startingStates <- c(1, 0, 0)
transitionsMatrix <- aperm(array(data = c(0.8, 0.1, 0.1,
                                          0.0, 0.7, 0.3,
                                          0.0, 0.0, 1.0,

                                          0.7, 0.1, 0.2,
                                          0.0, 0.6, 0.4,
                                          0.0, 0.0, 1.0,

                                          0.6, 0.1, 0.3,
                                          0.0, 0.5, 0.5,
                                          0.0, 0.0, 1.0,

                                          0.5, 0.1, 0.4,
                                          0.0, 0.4, 0.6,
                                          0.0, 0.0, 1.0,

                                          0.4, 0.1, 0.5,
                                          0.0, 0.3, 0.7,
                                          0.0, 0.0, 1.0),
                                 dim = c(3, 3, 5)),
                           c(2, 1, 3))
runmarkov(transitionsMatrix, startingStates, stateNames = states)


# Two state, four cycles, non-stationary model, state names not specified,
# 50/50 split between states at t=0

startingStates <- c(0.5, 0.5)
transitionsMatrix <- aperm(array(data = c(0.8, 0.2,
                                          0.3, 0.7,

                                          1, 0,
                                          0, 1,

                                          0, 1,
                                          0.5, 0.5),
                                 dim = c(2, 2, 3)),
                            c(2, 1, 3))
runmarkov(transitionsMatrix, startingStates)
}
